@@ -11,17 +11,24 @@
     }
   ],
   "require": {
-    "php": ">=7.0",
-    "kdyby/doctrine": "^3.1",
-    "kdyby/events": "^3.0",
-    "lookyman/nette-oauth2-server": "^2.0",
-    "league/oauth2-server": "^5.1,<5.1.4",
-    "nette/application": "^2.4",
-    "nette/di": "^2.4.1"
+    "php": "^7.1",
+    "kdyby/doctrine": "^3.3",
+    "kdyby/events": "^3.1, >=3.1.2",
+    "league/oauth2-server": "^7.0",
+    "nette/application": "^2.4, >=2.4.9",
+    "nette/di": "^2.4, >=2.4.10",
+    "lookyman/nette-oauth2-server": "^3.0",
+    "nette/caching": "^2.5, >=2.5.6"
   },
   "require-dev": {
-    "nette/bootstrap": "^2.4",
-    "phpunit/phpunit": "^5.4"
+    "nette/bootstrap": "^2.4, >=2.4.5",
+    "phpunit/phpunit": "^7.0",
+    "phpstan/phpstan": "^0.9.2",
+    "lookyman/coding-standard": "^0.1.0",
+    "phpstan/phpstan-nette": "^0.9.0",
+    "phpstan/phpstan-phpunit": "^0.9.4",
+    "phpstan/phpstan-strict-rules": "^0.9.0",
+    "jakub-onderka/php-parallel-lint": "^1.0"
   },
   "autoload": {
     "psr-4": {
@@ -32,5 +39,12 @@
     "psr-4": {
       "Lookyman\\NetteOAuth2Server\\Storage\\Doctrine\\Tests\\": "tests/"
     }
+  },
+  "scripts": {
+    "lint": "parallel-lint src tests",
+    "cs": "phpcs --extensions=php --encoding=utf-8 -sp --ignore=*/tests/temp/* src tests",
+    "tests": "phpunit --coverage-text",
+    "stan": "phpstan analyse -c phpstan.neon -l 5 src tests",
+    "check": ["@lint", "@cs", "@tests", "@stan"]
   }
 }
