@@ -7,85 +7,115 @@ use Psr\Log\LoggerInterface;
 
 class LoggerMock implements LoggerInterface
 {
+
 	/**
+	 * System is unusable.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function emergency($message, array $context = array())
+	public function emergency($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Action must be taken immediately.
+	 *
+	 * Example: Entire website down, database unavailable, etc. This should
+	 * trigger the SMS alerts and wake you up.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function alert($message, array $context = array())
+	public function alert($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Critical conditions.
+	 *
+	 * Example: Application component unavailable, unexpected exception.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function critical($message, array $context = array())
+	public function critical($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Runtime errors that do not require immediate action but should typically
+	 * be logged and monitored.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function error($message, array $context = array())
+	public function error($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Exceptional occurrences that are not errors.
+	 *
+	 * Example: Use of deprecated APIs, poor use of an API, undesirable things
+	 * that are not necessarily wrong.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function warning($message, array $context = array())
+	public function warning($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Normal but significant events.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function notice($message, array $context = array())
+	public function notice($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Interesting events.
+	 *
+	 * Example: User logs in, SQL logs.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function info($message, array $context = array())
+	public function info($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Detailed debug information.
+	 *
 	 * @param string $message
 	 * @param array $context
-	 * @return null
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function debug($message, array $context = array())
+	public function debug($message, array $context = []): void
 	{
 	}
 
 	/**
+	 * Logs with an arbitrary level.
+	 *
 	 * @param mixed $level
 	 * @param string $message
 	 * @param array $context
-	 * @return null
 	 */
-	public function log($level, $message, array $context = array())
+	public function log($level, $message, array $context = []): void
 	{
 	}
+
 }
