@@ -5,31 +5,28 @@ namespace Lookyman\NetteOAuth2Server\Storage\Doctrine;
 
 use Kdyby\Doctrine\Registry;
 use League\OAuth2\Server\RequestTypes\AuthorizationRequest;
+use Lookyman\NetteOAuth2Server\Storage\Doctrine\Client\ClientEntity;
 use Lookyman\NetteOAuth2Server\Storage\IAuthorizationRequestSerializer;
 
 class AuthorizationRequestSerializer implements IAuthorizationRequestSerializer
 {
+
 	/**
 	 * @var Registry
 	 */
 	private $registry;
 
-	/**
-	 * @param Registry $registry
-	 */
 	public function __construct(Registry $registry)
 	{
 		$this->registry = $registry;
 	}
 
-	/**
-	 * @param AuthorizationRequest $authorizationRequest
-	 * @return string
-	 */
 	public function serialize(AuthorizationRequest $authorizationRequest): string
 	{
 		$manager = $this->registry->getManager();
-		if ($authorizationRequest->getClient()) {
+		/** @var ClientEntity|null $client */
+		$client = $authorizationRequest->getClient();
+		if ($client !== null) {
 			$manager->detach($authorizationRequest->getClient());
 		}
 		foreach ($authorizationRequest->getScopes() as $scope) {
@@ -38,17 +35,17 @@ class AuthorizationRequestSerializer implements IAuthorizationRequestSerializer
 		return serialize($authorizationRequest);
 	}
 
-	/**
-	 * @param string $data
-	 * @return AuthorizationRequest
-	 */
 	public function unserialize(string $data): AuthorizationRequest
 	{
 		$manager = $this->registry->getManager();
 		/** @var AuthorizationRequest $authorizationRequest */
 		$authorizationRequest = unserialize($data);
-		if ($client = $authorizationRequest->getClient()) {
-			$authorizationRequest->setClient($manager->merge($client));
+		/** @var ClientEntity|null $client */
+		$client = $authorizationRequest->getClient();
+		if ($client !== null) {
+			/** @var ClientEntity $client */
+			$client = $manager->merge($client);
+			$authorizationRequest->setClient($client);
 		}
 		$scopes = [];
 		foreach ($authorizationRequest->getScopes() as $scope) {
@@ -57,4 +54,5 @@ class AuthorizationRequestSerializer implements IAuthorizationRequestSerializer
 		$authorizationRequest->setScopes($scopes);
 		return $authorizationRequest;
 	}
+
 }
