@@ -3,13 +3,12 @@ declare(strict_types=1);
 
 namespace Lookyman\NetteOAuth2Server\Storage\Doctrine\Client;
 
-use Kdyby\Doctrine\InvalidStateException;
-use Kdyby\Doctrine\QueryException;
 use Kdyby\Doctrine\Registry;
 use League\OAuth2\Server\Repositories\ClientRepositoryInterface;
 
 class ClientRepository implements ClientRepositoryInterface
 {
+
 	/**
 	 * @var Registry
 	 */
@@ -20,30 +19,26 @@ class ClientRepository implements ClientRepositoryInterface
 	 */
 	private $secretValidator;
 
-	/**
-	 * @param Registry $registry
-	 * @param callable|null $secretValidator
-	 */
-	public function __construct(Registry $registry, callable $secretValidator = null)
+	public function __construct(Registry $registry, ?callable $secretValidator = null)
 	{
 		$this->registry = $registry;
-		$this->secretValidator = $secretValidator ?: function ($expected, $actual) { return hash_equals($expected, $actual); };
+		$this->secretValidator = $secretValidator ?: function ($expected, $actual) {
+			return hash_equals($expected, $actual);
+		};
 	}
 
 	/**
 	 * @param string $clientIdentifier
-	 * @param string $grantType
+	 * @param string|null $grantType
 	 * @param string|null $clientSecret
 	 * @param bool $mustValidateSecret
-	 * @return ClientEntity|null
-	 * @throws InvalidStateException
-	 * @throws QueryException
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function getClientEntity($clientIdentifier, $grantType, $clientSecret = null, $mustValidateSecret = true)
+	public function getClientEntity($clientIdentifier, $grantType = null, $clientSecret = null, $mustValidateSecret = true): ?ClientEntity
 	{
 		/** @var ClientEntity|null $clientEntity */
 		$clientEntity = $this->registry->getManager()->getRepository(ClientEntity::class)->fetchOne($this->createQuery()->byIdentifier($clientIdentifier));
-		return $clientEntity
+		return $clientEntity !== null
 			&& $mustValidateSecret
 			&& $clientEntity->getSecret() !== null
 			&& !call_user_func($this->secretValidator, $clientEntity->getSecret(), $clientSecret)
@@ -51,11 +46,9 @@ class ClientRepository implements ClientRepositoryInterface
 			: $clientEntity;
 	}
 
-	/**
-	 * @return ClientQuery
-	 */
 	protected function createQuery(): ClientQuery
 	{
 		return new ClientQuery();
 	}
+
 }
