@@ -3,43 +3,29 @@ declare(strict_types=1);
 
 namespace Lookyman\NetteOAuth2Server\Storage\Doctrine\RefreshToken;
 
-use Doctrine\ORM\OptimisticLockException;
-use Doctrine\ORM\ORMInvalidArgumentException;
-use Kdyby\Doctrine\InvalidStateException;
-use Kdyby\Doctrine\QueryException;
 use Kdyby\Doctrine\Registry;
 use League\OAuth2\Server\Entities\RefreshTokenEntityInterface;
 use League\OAuth2\Server\Repositories\RefreshTokenRepositoryInterface;
 
 class RefreshTokenRepository implements RefreshTokenRepositoryInterface
 {
+
 	/**
 	 * @var Registry
 	 */
 	private $registry;
 
-	/**
-	 * @param Registry $registry
-	 */
 	public function __construct(Registry $registry)
 	{
 		$this->registry = $registry;
 	}
 
-	/**
-	 * @return RefreshTokenEntity
-	 */
-	public function getNewRefreshToken()
+	public function getNewRefreshToken(): RefreshTokenEntity
 	{
 		return new RefreshTokenEntity();
 	}
 
-	/**
-	 * @param RefreshTokenEntityInterface $refreshTokenEntity
-	 * @throws ORMInvalidArgumentException
-	 * @throws OptimisticLockException
-	 */
-	public function persistNewRefreshToken(RefreshTokenEntityInterface $refreshTokenEntity)
+	public function persistNewRefreshToken(RefreshTokenEntityInterface $refreshTokenEntity): void
 	{
 		if ($refreshTokenEntity instanceof RefreshTokenEntity) {
 			$manager = $this->registry->getManager();
@@ -50,15 +36,14 @@ class RefreshTokenRepository implements RefreshTokenRepositoryInterface
 
 	/**
 	 * @param string $tokenId
-	 * @throws InvalidStateException
-	 * @throws QueryException
-	 * @throws OptimisticLockException
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function revokeRefreshToken($tokenId)
+	public function revokeRefreshToken($tokenId): void
 	{
 		$manager = $this->registry->getManager();
 		/** @var RefreshTokenEntity|null $refreshTokenEntity */
-		if ($refreshTokenEntity = $manager->getRepository(RefreshTokenEntity::class)->fetchOne($this->createQuery()->byIdentifier($tokenId))) {
+		$refreshTokenEntity = $manager->getRepository(RefreshTokenEntity::class)->fetchOne($this->createQuery()->byIdentifier($tokenId));
+		if ($refreshTokenEntity !== null) {
 			$refreshTokenEntity->setRevoked(true);
 			$manager->flush();
 		}
@@ -66,22 +51,18 @@ class RefreshTokenRepository implements RefreshTokenRepositoryInterface
 
 	/**
 	 * @param string $tokenId
-	 * @return bool
-	 * @throws InvalidStateException
-	 * @throws QueryException
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function isRefreshTokenRevoked($tokenId)
+	public function isRefreshTokenRevoked($tokenId): bool
 	{
-		/** @var RefreshTokenEntity $refreshTokenEntity */
+		/** @var RefreshTokenEntity|null $refreshTokenEntity */
 		$refreshTokenEntity = $this->registry->getManager()->getRepository(RefreshTokenEntity::class)->fetchOne($this->createQuery()->byIdentifier($tokenId));
-		return $refreshTokenEntity ? $refreshTokenEntity->isRevoked() : true;
+		return $refreshTokenEntity !== null ? $refreshTokenEntity->isRevoked() : true;
 	}
 
-	/**
-	 * @return RefreshTokenQuery
-	 */
 	protected function createQuery(): RefreshTokenQuery
 	{
 		return new RefreshTokenQuery();
 	}
+
 }
