@@ -3,14 +3,13 @@ declare(strict_types=1);
 
 namespace Lookyman\NetteOAuth2Server\Storage\Doctrine\Scope;
 
-use Kdyby\Doctrine\InvalidStateException;
-use Kdyby\Doctrine\QueryException;
 use Kdyby\Doctrine\Registry;
 use League\OAuth2\Server\Entities\ClientEntityInterface;
 use League\OAuth2\Server\Repositories\ScopeRepositoryInterface;
 
 class ScopeRepository implements ScopeRepositoryInterface
 {
+
 	/**
 	 * @var Registry
 	 */
@@ -21,49 +20,44 @@ class ScopeRepository implements ScopeRepositoryInterface
 	 */
 	private $scopeFinalizer;
 
-	/**
-	 * @param Registry $registry
-	 * @param callable|null $scopeFinalizer
-	 */
-	public function __construct(Registry $registry, callable $scopeFinalizer = null)
+	public function __construct(Registry $registry, ?callable $scopeFinalizer = null)
 	{
 		$this->registry = $registry;
-		$this->scopeFinalizer = $scopeFinalizer ?: function (array $scopes) { return $scopes; };
+		$this->scopeFinalizer = $scopeFinalizer ?: function (array $scopes): array {
+			return $scopes;
+		};
 	}
 
 	/**
 	 * @param string $identifier
-	 * @return ScopeEntity|null
-	 * @throws InvalidStateException
-	 * @throws QueryException
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
-	public function getScopeEntityByIdentifier($identifier)
+	public function getScopeEntityByIdentifier($identifier): ?ScopeEntity
 	{
-		return $this->registry->getManager()->getRepository(ScopeEntity::class)->fetchOne($this->createQuery()->byIdentifier($identifier));
+		/** @var ScopeEntity $entity */
+		$entity = $this->registry->getManager()->getRepository(ScopeEntity::class)->fetchOne($this->createQuery()->byIdentifier($identifier));
+		return $entity;
 	}
 
 	/**
 	 * @param ScopeEntity[] $scopes
 	 * @param string $grantType
-	 * @param ClientEntityInterface $clientEntity
 	 * @param string|null $userIdentifier
 	 * @return ScopeEntity[]
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
 	 */
 	public function finalizeScopes(
 		array $scopes,
 		$grantType,
 		ClientEntityInterface $clientEntity,
 		$userIdentifier = null
-	)
-	{
+	): array {
 		return call_user_func($this->scopeFinalizer, $scopes, $grantType, $clientEntity, $userIdentifier);
 	}
 
-	/**
-	 * @return ScopeQuery
-	 */
 	protected function createQuery(): ScopeQuery
 	{
 		return new ScopeQuery();
 	}
+
 }
