@@ -9,34 +9,26 @@ use Lookyman\NetteOAuth2Server\Storage\Doctrine\Scope\ScopeRepository;
 
 class ScopeRepositoryMock extends ScopeRepository
 {
+
 	/**
 	 * @var ScopeQuery
 	 */
 	private $query;
 
-	/**
-	 * @param ScopeQuery $query
-	 * @param Registry $registry
-	 */
 	public function __construct(ScopeQuery $query, Registry $registry)
 	{
 		parent::__construct($registry);
 		$this->query = $query;
 	}
 
-	/**
-	 * @return ScopeQuery
-	 */
 	protected function createQuery(): ScopeQuery
 	{
 		return $this->query;
 	}
 
-	/**
-	 * @return ScopeQuery
-	 */
 	public function createQueryOriginal(): ScopeQuery
 	{
 		return parent::createQuery();
 	}
+
 }
